requirements.txt
streamlit
tensorflow
matplotlib
numpy
yfinance
plotly
newsapi-python
textblob

joblib

